refactor(autopay): extract AutopayModalBody comp
